+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 ## v1.0 [![27.05.2020](https://img.shields.io/date/1590589020)](https://github.com/d8corp/react-mobx-routing/tree/v1.0)
+- `rel="noreferrer"` and `target="_blank"` by default for external links
 ### v1.0.8 [![03.06.2020](https://img.shields.io/date/1591133141)](https://github.com/d8corp/react-mobx-routing/tree/v1.0.7)
 - fixed dependencies
 ### v1.0.7 [![02.06.2020](https://img.shields.io/date/1591130194)](https://github.com/d8corp/react-mobx-routing/tree/v1.0.7)